Home page touch-ups
Cleaned up the home page. Took screenshots of old home page and stored in folder called "Screenshots" for comparison and presentation. Altered the colors of the backgrounds, removed white borders to smooth the look (kept color-scheme to oranges, yellows and blues).

Created a logo (subject to change if others do not like). Fixed the "pellio" logo so no white shows.

Altered the login and "new account?" buttons to be more visually aesthetic.

FIXED THE RADIO BUTTONS WOOHOO. Now it only allows you to select one option. Also deleted some of the "feelings" because I felt some were redundant but we can add them in if needed.